+++ src/platform_impl/macos/window.rs
@@ -79,6 +79,8 @@ pub struct PlatformSpecificWindowBuilderAttributes {
     pub disallow_hidpi: bool,
     pub has_shadow: bool,
     pub accepts_first_mouse: bool,
+    pub allows_automatic_window_tabbing: bool,
+    pub tabbing_identifier: Option<String>,
 }
 
 impl Default for PlatformSpecificWindowBuilderAttributes {
@@ -94,6 +96,8 @@ impl Default for PlatformSpecificWindowBuilderAttributes {
             disallow_hidpi: false,
             has_shadow: true,
             accepts_first_mouse: true,
+            allows_automatic_window_tabbing: true,
+            tabbing_identifier: None,
         }
     }
 }
@@ -325,6 +329,14 @@ impl WinitWindow {
                 }
                 if pl_attrs.movable_by_window_background {
                     this.setMovableByWindowBackground(true);
+                }
+
+                if pl_attrs.allows_automatic_window_tabbing {
+                    this.setAllowsAutomaticWindowTabbing(true);
+                }
+
+                if let Some(identifier) = pl_attrs.tabbing_identifier {
+                    this.setTabbingIdentifier(&NSString::from_str(&identifier));
                 }
 
                 if attrs.always_on_top {
@@ -1215,6 +1227,26 @@ impl WindowExtMacOS for WinitWindow {
     #[inline]
     fn set_has_shadow(&self, has_shadow: bool) {
         self.setHasShadow(has_shadow)
+    }
+
+    #[inline]
+    fn set_allows_automatic_window_tabbing(&self, enabled: bool) {
+        self.setAllowsAutomaticWindowTabbing(enabled)
+    }
+
+    #[inline]
+    fn allows_automatic_window_tabbing(&self) -> bool {
+        self.allowsAutomaticWindowTabbing()
+    }
+
+    #[inline]
+    fn set_tabbing_identifier(&self, identifier: &str) {
+        self.setTabbingIdentifier(&NSString::from_str(identifier))
+    }
+
+    #[inline]
+    fn tabbing_identifier(&self) -> String {
+        self.tabbingIdentifier().to_string()
     }
 }